requirements.txt
python-telegram-bot
Pillow
pytesseract